starting visualization refactor #258
